requirements.txt
bcrypt==3.1.2
cffi==1.9.1
Django==1.10.5
django-extensions==1.7.6
gunicorn==19.6.0
pycparser==2.17
python-dateutil==2.6.0
pytz==2016.10
RandomWords==0.2.0
six==1.10.0
